Reorganizing Codebase Without Disrupting Functionality
- [+] feat(terminal): add retry policy for executing RetryableFunc with exponential backoff
- [+] fix(terminal): remove duplicate retryWithExponentialBackoff function from session.go
