+++ terminal/session.go
@@ -7,20 +7,15 @@ import (
 	"bufio"
 	"context"
 	"fmt"
-	"math"
 	"os"
 	"os/signal"
 	"strings"
 	"sync"
 	"syscall"
-	"time"
 
 	"github.com/google/generative-ai-go/genai"
 	"google.golang.org/api/option"
 )
-
-// RetrypolicyFunc is a type that represents a function that can be retried.
-type RetryableFunc func() (bool, error)
 
 // Session encapsulates the state and functionality for a chat session with a generative AI model.
 // It holds the AI client, chat history, and context for managing the session lifecycle.
@@ -37,30 +32,6 @@ type Session struct {
 	// for example, if session is ended not cause of client, then it will be renew with previous chat history.
 	lastInput string // Stores the last user input for reference
 
-}
-
-// retryWithExponentialBackoff attempts to execute a RetryableFunc with a retry policy.
-func retryWithExponentialBackoff(retryFunc RetryableFunc) (bool, error) {
-	const maxRetries = 3
-	baseDelay := time.Second
-
-	for attempt := 0; attempt < maxRetries; attempt++ {
-		success, err := retryFunc()
-		if err == nil {
-			return success, nil
-		}
-
-		if logger.HandleGoogleAPIError(err) {
-			delay := baseDelay * time.Duration(math.Pow(2, float64(attempt)))
-			time.Sleep(delay)
-			continue // Retry the request
-		} else {
-			// Non-retryable error or max retries exceeded
-			return false, err
-		}
-	}
-
-	return false, fmt.Errorf(ErrorLowLevelMaximumRetries)
 }
 
 // NewSession creates a new chat session with the provided API key for authentication.
